Send 403 if contact does not belong to user

+++ delete.php
@@ -1,14 +1,16 @@
 <?php
     require "database.php";
-    $id = $_GET["id"];
 
     session_start();
-    session_start();
+
+    
     if (!isset($_SESSION["user"])) {
       header("Location: login.php");
       return;
     }
     
+    $id = $_GET["id"];
+
     $statement = $conn->prepare("SELECT * FROM contacts WHERE id = :id");
     $statement->bindParam(":id", $id);
     $statement->execute();
@@ -17,6 +19,14 @@
         http_response_code(404);
         echo ("HTTP 404 NOT FOUND");
         return;
+    }
+
+    $contact = $statement->fetch(PDO::FETCH_ASSOC);
+
+    if ($contact["user_id"] != $_SESSION["user"]["id"]){
+      http_response_code(403);
+      echo ("HTTP 403 UNAUTHORIZED");
+      return;
     }
 
     $conn->prepare("DELETE FROM contacts WHERE id = :id")->execute([":id" => $id]);
